pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neuralbridge"
version = "0.1.0"
description = "Universal Enterprise Middleware for Agentic AI Integration — securely connect ANY AI agent to ANY system with YAML configuration, CRA compliance, and zero-trust security."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "NeuralBridge Contributors"}
]
keywords = [
    "agentic-ai", "ai-agent", "middleware", "enterprise",
    "openclaw", "mcp-gateway", "cra-compliance",
    "zero-trust", "adapters", "langchain", "autogpt"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: System :: Systems Administration",
    "Framework :: FastAPI",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "cryptography>=42.0.0",
    "pyjwt>=2.8.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "structlog>=24.1.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "prometheus-client>=0.20.0",
    "tiktoken>=0.6.0",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "httpx>=0.27.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/neuralbridge/neuralbridge"
Documentation = "https://neuralbridge.readthedocs.io"
Repository = "https://github.com/neuralbridge/neuralbridge"
Issues = "https://github.com/neuralbridge/neuralbridge/issues"

[project.scripts]
neuralbridge = "neuralbridge.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "DTZ", "T20", "ICN"]
ignore = [
    "E501",   # Line too long — handled by formatter
    "S101",   # Use of assert (used for type narrowing)
    "S104",   # Possible binding to all interfaces
    "S105",   # Possible hardcoded password
    "S311",   # Standard pseudo-random generators
    "S608",   # Possible SQL injection (parameterised in practice)
    "DTZ001", # datetime.now() without tz (used in Pydantic defaults)
    "DTZ003", # datetime.utcnow() (used in Pydantic defaults)
    "B008",   # Function call in default argument (FastAPI Depends)
    "A002",   # Shadowing built-in (e.g. license)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=neuralbridge --cov-report=term-missing --cov-report=html"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
